Even stricter scheduler assertions
Also: Thanks to these assertions, I've finally managed to find the bug for which I've added them. - The problem were the IPI functions in rpc-async not checking for `target == THIS_TASK` when re-scheduling the freshly awoken thread to be executed as soon as possible, causing scheduler inconsistencies after the calling thread accidentally unlinked itself.