Fix zone-specific accounting of the free-pages counter
I took forever to find this one, even going so far as to add a new out-of-kernel (as in: driven by magic) sub-system for tracing physical memory allocations beacause I was convinced that there was still some part of the kernel that leaked memory during the fork()->exec()->exit() cycle of program execution.
But it turns out that even though /proc/kos/raminfo kept on indicating that more and more memory was being used, that was actually a lie, and all that was wrong was `page_freeone()' not incrementing the free-pages counter of the associated zone...
... What a doozy. - It's not often you find a (quote-on-quote) ~leak~ that's caused by incorrect accounting of allocations.

Anyways. - Repeatedly executing `cat /proc/kos/raminfo' will now present you with a fairly consistent value for `used' (it only goes up/down on occasion, relative to various caches being filled and empied)