Fix a bug in atree:minmaxlocate_at()
-> This also fixes deemon being unable to load dex modules!
In case you're interested, this one caused mprotect() to fail (but even had the ability to panic() the kernel under certain circumstances).
In this case however, it surfaced by causing deemon to be unable to load dex modules because they had the TEXTREL flag set, which then used mprotect(), which then didn't work properly, which then caused a SEGFAULT when libdl tried to apply R_386_RELATIVE relocations. -> It was _that_ _simpe_ ;) (emphasis on the ~simple~)