Start implementing USB support for isochronous and interrupt-driven devices
It's not finished, but the software interface design is complete, and most of the implementation is finished as well. (The idea is that you register interrupt handlers with a USB controller, where you're able to specify the interval in which the controller will poll your endpoint, where 0 means constantly for interrupt-driven devices, and any other value being used for isochronous interrupts)

This functionality will be required to implement interrupt-driven (e.g. keyboard), as well as isochronous (e.g. microphone) USB devices