Discover a flaw in push segment registers are saved
This affects how libunwind, as well as a couple of other places access saved segment registers values, which is currently done in a way that is incompatible with what Intel documents as the actual behavior on certain older processors (though not all processors).
Note that this commit isn't a fix, but rather the addition of TODOs, as well as a test program for reproducing the problems caused by this.