New way to get SHF_ALLOC for .debug_info
This new method is much more robust because it doesn't rely on an oversight in the programming of `ld' any more, and also works to allocate compressed .debug_info sections.
Note however that libdebuginfo doesn't yet support compressed debug info sections, and that support for such a feature is planned to be added in the future. (and there's a good reason to support compression, which you can read about in the comment added to `kos/.sources')

Also: This new method of doing it is somewhat cleaner since it no longer relies on specially formatted data descriptors to be injected into the debug sections in order to make them be allocated, but instead manually edits the kernel binary post-linking in order to re-arrange program headers, allowing debug information to become mapped.