Try to implement by own inflate function
It's working, but unless you're decompressing your entire data-blob at once, it's unbearably slow when trying to decompress only bits and pieces at a time.
The idea is to use this to have BFD compress .debug_* sections at compile-time, and decompress them on-the-fly, whilst inside of the builtin debugger.
That way, the kernel binary could become much  smaller, whilst still retaining the same debugging functionality

I am aware that I can implement a cache for huffman trees, but the big slowdown in decompression performance comes from the data-repeat compression method that requires knowledge of up to 32K bytes (sliding window) of already decompressed data, immediatly preceding the current byte, which is something that I would like to work around. And even still, since I would like to have some kind of interface for simply reading data from decompressed .debug_* sections at arbitrary offsets, the sliding window would have to be re-calculated each time by once again decompressing _every_ _single_ byte of data preceding the current location (so that's O(position), meaning that the function would (and does) become unbearably slow over time), so a sliding window wouldn't help either when it comes to reading from arbitrary offsets.
The only solution that I can see at this point would be to have the debugger decompress all debug data ahead of time, but that would mean that the debugger would no longer work when there is insifficient memory available (and I'm talking _a_ _lot_ of required memory here, as in the ranges of more than 6MiB)

The last solution that comes to mind would be to hack BFD's (that the back-end for `ld' by the way), to not make use of a sliding window, and not generate distance/length symbol pairs during compression, however that seems sketchy at best, and would pretty much ruin its great compression ratio of reducing debug sections to only about 1/10th of their size (which it only does because the sections contain so much repeating data)