Introduce return type annotations in magic headers
With this, [nonnull] can be used with return types, and attributes of malloc-like functions can be automated much better than they could be before.

Note however that only a small hand full of functions are already making use of these new annotations, but that number will probably increase in the near future.