Add a new field `e_faultaddr' to `struct exception_data'
This field holds the original program counter of the instruction that caused a fault to happen (or the same address as already stored in the exception-register-state in case of a trap).
Note that the pc stored in the exception-register-state _always_ points after the instruction that caused an error (even in the case of something like a #UD of #PF)
With the introduction of this field, it is now possible to safely determine the address of a faulting instruction, without having to use `instruction_pred()' and hope that nothing has changed since the error, meaning that this use-case has now become much more robust!