Completely re-write the system call routing system
With this change, system call meta-data is generated much more efficiently, and many aspects related to system calls that used to be arch-specific are now generic.
The main change implemented is to split the definition and binding of system calls, as well as remove all arch-specific code from the system call generator script.