Re-implement the x86_64 qemu multiboot work-around
With this, you're finally able to boot the x86_64 kernel, and ... just wow!
I would not have believed that all of my bootstart assembly would function out-of-the-box. But, as it turns out: All of the startup checks, CPU mode transitions, paging initialization, and all of that funny stuff already works! Even GDB debugging already works without me having to do anything else! (which is why I exactly know where the kernel currently crashes, because come on: it wasn't like the thing would be able to reach user-space without some bugs), but still: holy $h1t! I was not expecting the first ever start-up to get that into the kernel initialization.
Right now, the kernel breaks in `kernel_initialize_boot_trampolines()', meaning that this is where the next game-breaking bug lies

Yay! x86_64 support! (kind-of)  \(^w^)/