Provide a new API for paging without vm_vpage_t/vm_ppage_t
The idea is to eventually remove these 2 types all-together, and get rid of the differentiation between page-indices and addresses. The end-goal here is that as far as paging conventions go, there should only be 3 different types of addresses: Virtual (`void *'), Physical (`vm_phys_t'; to-be renamed to `phys_t') and File (`pos_t'). Additionally, `size_t num_pages' should be replaced with `size_t num_bytes'.
Any other kind of indexing mechanism should no longer appear in any public kernel API, and any API that one of these types does exist in should be considered deprecated/pending-to-be-changed.
The idea of this entire change is to break out of the corner that I wrote myself into in regards to the sign-extension of virtual pointers under x86_64, which could have been fixed by splitting apart the `VM_ADDR2PAGE' family of functions, however thinking about the whole thing, the entire idea of constantly converting back and forth between page indices and actuall addresses is just entirely redundant to begin with, and only exited up until this point because it ~was~ part of the well-established KOS arch-API.