Fix a soft-lock caused by not closing buffers during destruction
They _always_ have to be closed because some thread may still be connected. And while the signal system does allow unprocessed async signals to already have been destroyed, those signals still have to be broadcast prior to being destroyed in order to prevent threads that are still waiting on them to lock up.