Design a kernel interface to facilitate `posix_spawn()'
Also: This interface includes a couple of extensions to fix some missing features in `posix_spawn()' (such as the ability to create a new process in a new process group, and then have that process group become the foreground group of a custom terminal; this functionality will likely be exposed via an extension `int posix_spawn_file_actions_addtcsetpgrp_np(posix_spawn_file_actions_t *file_actions, fd_t ttyfd);' that has yet to be designed)
NOTE: This is _not_ the implementation of such an interface! This commit only adds prototypes, structures and flags (i.e. a layout of the planned API/ABI). None of these functions have yet to be implemented.