Design a user-space API for async I/O
Not implemented yet (and may still be changed), however this reflects the general layout in which AIO will be made available to user-space.
Note that this API allows for a not-implemented completion status! aka: Kernel objects aren't required to implement whatever will eventually take the place of the backing API