Add support for a permanent physical memory identity mapping on x86_64
The first 64TiB of physical memory can be identity mapped into virtual memory. Anything beyond this requires the use of trampoline memory mappings.
This is acomplished either through use of 1GiB mappings (if supported), or lazily initialized, and dynamically re-used 512*2MiB mappings (with a configurable upper limit on vector overhead)