Initial support for booting KOS using VirtualBox
It was quite hard to automate the processing to creating+configuring the KOS VM, as well as getting the the x86_syslog_port->emulator->stdout pipeline working (I had to dive into named pipes; s.a. the latest deemon commit), but the basic framework is now complete.
With all of that out of the way, KOS current ends up in an infinite loop within libvm86 (as driven by libemu86), so I'm guessing that there's something I'm doing wrong in the emulation of some instructions, but now that VirtualBox support is out of the way, I can actually start tackling this problem!