Lazily patch use of (rd|wr)(fs|gs)base in drivers on x86_64
With this, code in `handle_bad_usage.c' will automatically try to patch these instructions with function calls that emulate the behavior of these instructions, such that any possible slow-down from use of such instructions only happens once (namely: during the first time when the emulated call is injected).
In a way, this functions like lazy relocations, only that no descriptions are needed, and the patching itself happens entirely transparent to the individual drivers.