Completely re-do libinstrlen for i386/x86_64
The library now has options to pass the ISA of the instruction to analyze, as well as properly handle AMD XOP instructions.
Finally, the database used on i386/x86_64 is now based on libdisasm, meaning that it can be automatically updated to match whenever changes/additions are made to libdisasm!

Note however that users of the library have yet to be updated such that everyone is specifying the mandatory ISA codes. - That will happen next.