Don't use __builtin_nan() in the implementation of libc
Looking at the output of `readelf -r', it appears that __builtin_nan() _does_ call its library variant in (at least) some cases. As such, this variant cannot be used in order to implement the pure, low-level math implementations!
This commit fixes 3 unnecessary relocations appearing in both the 32-bit and 64-bit libc.so libraries.