Fix realtime() running to fast
Was caused by not accounting for quantum override during a quantum reset (as can happen as the result of calling `task_yield()'). Not considering this caused a quantum overrun to be detected the next time a call to `realtime()' was made, resulting in the realtime() clock being advanced by 1/20 of a second, leading to the clock appearing to running too fast, before being able to catch up once once the system eventually got settled.

With this, realtime() appears to be working flawlessly, and even tetris works without any random jitters, slow-downs, or block suddenly dropping across half the screen! :)