Generate _all_ headers using the new magic
Also: Fix some problems that broke compiling KOS due to recent changes in system headers. (magic completes without errors after touching "include-everything.h" and "src/libc/api.h", i.e. when doing a full re-build of the _verify system, as well as KOS's C library)

With this commit, the first milestone of automating the generation of system headers has (once again) been reached. (in that all headers such as <string.h> are being generated once again, and do a sufficiently well job to be usable as a replacement for the old magic from now on)
Also note that these regular headers are the most complicated part of the whole process, as they are where all of the complex function bindings take place in. None of the other files actually need to do anything that's more complicated that this in terms of the binding of functions