Fix a race condition in FunctionImplementation.dependencyFunctions
This used to have the potential of causing function dependencies to become lost within libc auto implementations, resulting function names not being escaped property, and later-stage compiler errors to happen.
The problem was that the property used to check for m_dependencyFunctions being bound, but fill in that member before having fully initialized it. If at any point during execution of the remainder of the property, some other thread were to call the property again, that thread would be given an incomplete Dict of deendencies, possibly still empty at that point in time. - With this, that case is now handled by initializing the field multiple times, which is what was intended from the start, and doesn't have any risks associated with it