Re-write the preprocessor optimizer
With this, the optimizer is able to operate as context-dependent, as well as being able to optimize a larger range of expressions.

I'm extremly happy with how this turned out, and you can look through the changes made by this commit to generated files to convince yourself of how useful these new features are for removing completely unnecessary bloat from generated files.

Essentially, this allows the code generator to optimize expressions based on assumptions gathered from surrounding conditional blocks. For example: any code inside of '#ifdef foo' is allowed to assume that 'defined(foo) == 0'.
Additionally, the optimizer is now also able to recognize stuff like `4 == 2' (as produced by uchar-replacements) as a constant expression