New KOS-specific ELF extension: STT_KOS_IDATA
This is a custom ELF symbol type that behaves like STT_GNU_IFUNC, but instead is used for data-objects. With it, lazy initialization and address resolution of global variables becomes possible, the same way this can be done with resolver functions.
Using this, KOS's libc is now able to export `sys_errlist' (which is a vector of string-pointers, and as such would normally require the use of ~evil~ relocations)
But with this extension, libc can provide a resolver funciton for `sys_errlist' that will lazily initialize the vector at runtime, but will _only_ do so when the hosted application is actually using that variable (or the first time that someone calls `dlsym(RTLD_DEFAULT, "sys_errlist")').