Implement support for RTM memory far regions
With this, VM data blocks can be loaded into RTM caches in segments, rather than having to be loaded as consecutive memory ranges.
As a result of this, `playground rtm' now works without having to set `LD_BIND_NOW=1', meaning that the entirety of libdl's lazy symbol binding can now be executed in the an RTM context!

wow: that's probably the most amount of consecutive code that libemu86 had to execute all at once to date. - I've counted it, and to service the RTM block from playground, including the lazy-symbol-binding part, it takes a total of 1671 instructions that have to be emulated. (and in the end, it actually works!)

Next up: Preprocess-away a whole bunch of the printk-trace statements within the RTM driver...