Fix a bug with 8-bit register emulation in RTM-mode
Of all the things, this caused the control-character switch in libdl's format_printf() to break and always branch to the default-case, causing the lazy-resolve syslog entry to contain the raw format characters, rather than what they represent.
With this, that is now fixed, and `playground rtm' correctly logs '[rtld] Lazy resolve "sys_rtm_end" in "/bin/playground" (to ... from "/lib64/libc.so")' to the system log after successful RTM completion.