Fix a problem with intellisense on x86_64
Fixed by prevending pre-defined macros, which is also done to nudge users towards always including headers such as <hybrid/typecore.h> if the intend is to use macros such as __SIZEOF_POINTER__ or the like. (since failure to do so will now result in an intellisense error)
By the way: The problem with intellisense on x86_64 was that __has_include() was pre-defined as a macro that broke its behavior in a way that couldn't actually be fixed without disabling _all_ builtin macros.