More work-arounds for the gcc c++-using bloat bug
So far, I've been able to cut out ~400KiB from the x86_64-OD kernel, which may not sound too much, but is still a considerable amount of data. Especially when considering that this the amount of compressed debug info I was able to cut out.
But there's still more left to debloat...