Correctly handle E_WOULDBLOCK during sysret
This can happen as the result of touching user-space memory, which is done as part of verifying that the return-PC is canonical.
Note that when this happens, it doesn't actually indicate that user-space did anything wrong, as this can simply happen when syscall returns to a text location that has yet to be been loaded from disk.