Add *support* for libdrm
While it can be built, I very much doubt that it would work in its current state. But for the time being, my plan is to port everything needed to get Xserver running under KOS (yea: I know there's a decent chance that KOS is nowhere even close to supporting some interface that I've neven even heard of, but is required to get Xserver running (unix domain sockets are probably one of those interfaces; but then again: At one point, KOSmk3 (or as is mk2?) already had support for unix domain sockets)), but irregardless of any of that, the current plan is to get an Xserver binary that can be ~run~ *started* under KOS, as well as some X application that tries to talk to it.
Once I've gotten that far, I'll try to implement whatever it'll be that's actually going to prevent exactly that from working. (small steps... And with KOS pretty much having mastered the 80s-style terminal world, it's time to head down the road of proper, modern Graphics/GUI applications, though I can't make any promises as to whether or not KOS will ever get there...)