Move video-mode switching code into its own library
This code is used by the builtin debugger, and must be able to switch to 80x25 text-mode from whatever configuration the display was in before, as well as be able to switch back when the debugger is exited.

The next step will be to make it so that the new libvgastate is able to switch away from extended video modes, as may have been enabled by the Xorg server.