Fix python config to _actually_ use the system libffi
Ugh... Does anyone even test this stuff? I mean: Why does python thing it has to re-invent the wheel when it comes to figuring out how to link against a system libffi, when $PKG_CONFIG could already tell it everything it needs. - Heck: It even uses $PKG_CONFIG, but then proceeds to only use it for figuring out the FFI include path, which if that one's empty (as is the case on KOS, since ffi gets installed to KOS's default #include path), then it completely mucks up.
And that isn't even mentioning that unless you were to patch its build script, or hide some files before running it (I'm doing the later), it will link against the static version of libffi (or rather: `libffi_convenience`, which only comes as a static version)
Anyways: It's working now, and produces a `_ctypes.so` that is linked against `libffi.so.7`, just how I want it to be.