Support unwind+addr2line in the debugger for user-space
With this, you can now use `trace' within the builtin debugger to generate tracebacks that include the underlying user-space applications.

It's still in the testing phase, and I do plan on adding a cache, since at the moment every addr2line operation includes a lot of overhead as the result of having to load, map, and umap a number of sections into kernel-space.