Fix a reference leak related to file synching
When more than one INode was modified, there was a situation where some of the modified nodes were wrongfully (and without decref) evicted from the chain of modified INodes.

With this fixed, I was able to add a no-system-leaks check to system-test, without that check breaking, meaning that (at least as far as system-test goes), the KOS kernel should now be leak-free!