Implement a new API for low-level signaling
The new API closely follows the old one, however as far as implementation details got, it greatly differs from it. This new implementation fixes some race conditions I've noticed related to sending signals while receiving threads are being destroyed (admittedly, this is something so rare that I've never even seen it happen once), as well as with a greater emphasis on ensuring that signals generated by `sig_send()' will try their hardest to find their way to some receiver thread, such that now they will be forwarded to other waiting threads when the original recipient decides to discard the signal (as is able to happen due to race conditions, which used to lead to a situation where a mutex's lock becomes available, but none of the threads actually interested got properly informed). Alongside this, the notion of ghost connections has been removed (which was a feature that never really found great use, as it's simply too difficult to properly determine when a connection should be a ghost, since this choice could only be made by code too high-level to still be able to fully understand the meaning of such a choice). (Ghost connections are the ones that should be triggered by `sig_send()', but not be considered a viable receiver) With this change, the signal API should be able to automatically determine when a connection should be what used to be a ghost, since unused single-target signals are now forwarded to other waiting threads.