Workaround for a gcc bug
gcc unconditionally (as in: no commandline option can prevent this) assumes that a function named `memcpy()' (along a long list of other functions) can never throw an exception, no matter how the function is declared, and no matter if you pass `-fnon-call-exceptions', `-ffreestanding' or `-fno-builtin'.
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=96985

I did however discover that it doesn't make this assumption when you declare the function as `extern "C++"', which is exactly what this commit introduces for all functions that gcc recognizes as noexcept/nothrow, but that aren't declared as NOTHROW by KOS system headers. (thank science that KOS system headers get generated automatically, so all I really had to do was add some macros, and some special handling to the declaration generator, so that functions like `memcpy()' now get declared as `extern "C++" ... memcpy(...) __asm__("memcpy");'. Because this would have been one hell of a lot of work to fix this by hand...)