Fix a scenario where pgrp-signals didn't reach everyone
It might sound impossibly rare: but there was a data-race where one thread got the signal, and terminated as a result, while another had yet to get the signal, fork()'ed a new thread, with that new thread's descriptor then got allocated at the same location in kernelspace as the previous thread that already exited. Since the `task_raisesignalprocessgroup()' (used to) simply track which processes already got a signal delivered via a hash-set based on the address of the individual thread descriptors, this kind of scenario meant that it would end up skipping a thread, thinking that thread already got the signal when in actuality that thread was created just a moment ago, and just by random chance took the same address as a now-dead thread.
Anyways: Even though only this commit includes it, if you were to look at other recent commits, I've been working on getting KOS to behave correct during fork-bombs (s.a. `playground forkbomb'), and this was the last puzzle-piece to make it possible to halt the bomb with something like CTRL+C