Implement `pthread_sigmask()' as an alias for `sigprocmask()'
On KOS, both of them only affect the calling thread, where POSIX claims that use of sigprocmask() is unspecified in a multithreaded process, while pthread_sigmask() only affects the caller. In other words: it should be OK to have both of them work for only the caller in a multithreaded process.