(WIP) Fix various problems with RTC/TSC timings
Mainly wrap my head around how to properly calculate a high-precision HZ value, such that precision increases over time, as well as how APIC interrupts function, and when one has to increment the APIC TSC base offset in relation to said interrupt.

The result of all of this work now is that `playground sleep' (ithat is: the program used to measure kernel timekeeping precision) is running more accurate than ever: After letting the kernel run for a while and starting the program, as well as pulling up my system clock in a different window, they actually roll over to the next second at pretty much the exact same moment! \(^v^)/

Still (WIP) because I know for certain that the i386 version of the kernel can't be built at the moment, since it's missing the implementation for uint128_div and uint128_mul. Once that's been re-added, and once the kernel can once again be built under all configurations, this (WIP) will go away