Account for base-line priority boosts during high-priority switches
When a low-priority thread switches to high-priority thread, it gifts the remainder of its quantum to that thread. However, in addition to this gift, the high-priority thread must also include its own, original priority boost (the one it gained by waiting for a while) when calculating its new deadline. - Else, the priority boost gained from waiting for a while would get lost.