Fix some race conditions related to threads moving between CPUs
When reading the t_cpu field of a thread, one always has to be prepared for the possibility that the thread may have changed CPUs just afterwards. The only exception to this is when thread is hosted by your own CPU, and you've disabled preemption before the read.

I've had this manifest as a dead-lock in task_wake(), where the thread to-be awoken changed CPUs between the initial read from its CPU field, and the point where I was sending the IPI to the target CPU. Becuase the t_cpu field was read multiple times, this lead to the scenario where the IPI path for used such that the caller of task_wake() tried to send an IPI to themself, which they weren't allowed to do with the flags passed to `cpu_sendipi()', leading the send-ipi loop to run indefinitely.