Completely re-write the debug-malloc system
The new system is designed around an idea I've had to ensure that one is always able to untrace any memory location, without the act of doing so possibly blocking, or eventually failing when too many calls have been made, such that the kernel can no longer keep up.

The solution is basically to re-use the traceback area of the internal tracing nodes as a bitset of word-locations still considered as traced in the event that an untrace operation would have to cut a hole into tracing-node.
And as far as locking goes, simply use a global preemption+SMP lock that is always held for as short a time as possible, thus satisfying the NOBLOCK requirements and making everything safe in regards to re-entrancy (such as calling kfree() from an interrupt handler, or async-RPC)