Use `atomic_lock' instead of `atomic_rwlock' when possible
The former can generate code that is more efficient than the later. The only reason why a lot of code had always been using atomic_rwlock when it didn't actually need write-locks was just because `struct atomic_lock' was a recent addition when I realized the potential performance-impact of the later in certain situations.