Design a new CFI extension to deal with landing-adjustments
This extension allows one to define capsules of CFI instrumentation that get executed both during normal unwinding, as well as when having to leave the ~casule~ as part of jumping to a local exception handler. (thus allowing one to specify certain actions that must _always_ be performed when exiting from a certain PC-range)
Note that normal CFI doesn't allow one to specify custom unwind behavior before jumping to a local exception handler (which is a problem already somewhat addressed by `DW_CFA_GNU_args_size', which even get generated by GCC itself)

Note this commit already provides the specs for the CFI capsule extension, as well as a new API for libunwind related to unwinding for the purpose of executing CFI code before jumping to a local exception handler. However, the code used to scan for CFI capsules is still missing...