Fix a flaw with how read-lock descriptors are managed by rwlock-s
The thing just kept on allocating more and more descriptors over time, never try to clear the hashmap in order to free up previously deleted entires.
Mind you that the system was allowed to do this (semantially speaking), and that it doing this didn't cause any stability issues. However continuously allocating more and more memory really isn't something that's intended (though I should note that when a thread is destroyed, this hashmap was still freed, and furthermore: memory consumption over time was linear, so this wasn't actually a major problem)