Implement support for printing structs in moddbx
With this, the basic cprinter API is pretty much complete. However, I do intend on adding functionality for custom extensions to how certain objects are printed (e.g. it'd be nice if `struct handle' would show the pointed-to object, or `struct task' would show what the task belongs to, etc...)