Add support for user-space TLS to moddbx
With this, moddbx is able to load the value of ELF TLS variables, such as libc's `current' (iow: what's returned by pthread_self()).
The implementation is insanely hacky though, and relies on the implementation of libdl's TLS backend (s.a. the macros that encoding the fixed offsets, which can be found in moddbx/cexpr.c