Add support for DW_OP_addr in shared libraries
It's not really stated by official documentation (DWARF5.pdf), but you have to add the load-offset of a shared library to DW_OP_addr when the instruction appears inside one of the non-allocated .debug_* sections. However, when it's from one of the allocated sections (such as .eh_frame), then that offset mustn't be added, and it is up to the module to emit a relocation for that location, and have the dynamic linker fix-up the address during startup.