Implement the pthread_key_t API
Note however that this API is anything but efficient. It's designed work and be correct, and not and _any_ additional static overhead to existing TLS structures (by using dltlsalloc() as its bakend).
However, this also means that it's quite slow, meaning that it is _very_ _much_ recommended you keep on using ATTR_THREAD if you want a variable to be thread-local.