Create a framework to implement DW_OP_entry_value
Note that by no means has support for that instruction already been implemented. - This commit simply moves the work required to implement it down one layer of abstraction, defining the KOS-specific abi for instruction executed in the context of DW_OP_entry_value, as well as defining an API with which one will (eventually) be able to do what is required by this opcode.
I've chosen to create a new library for this purpose, simply because doing this properly may take quite a lot of code that would otherwise just unnecessarily bloat libunwind.