Design a new system to simplify system headers
I admit it: KOS's system headers have gotten pretty bloated in terms of the raw size of certain headers. This however, I plan on changing, but not by removing features from headers, but rather by generating and using simplified versions of headers that are re-written to compile much faster, whilst only working for a specific compiler/arch-combination.
The system I've designed for this is called CimPPlify, and allows one.
If you want to talk stats: using the script to simplify <string.h> results in a a file that is ~70% smaller than the original