New (but unfinished) memory manager design
A lot of the low-level stuff has already been implemented, but I really wanted to commit the work already done, since it's already been a couple of non-stop work.

Note that this is pretty much a full re-write from scratch, and that the new code is of much greater quality than the old code.

But also note that the new code can't actually be compiled (yet) since many parts are not implemented yet, and (in some cases) havn't even been (re-)designed, yet.

In general, this re-write changes many of the low-level details of mman components, but keeps the general high-level approach. Probably the biggest change is that mpart (vm_datapart) no longer need to match the size of mnode (vm_node) exactly, but can be off by a bit, both in the front and the back. Another big change is that the (intentional) reference loop between mfile (vm_datablock) and mpart (vm_datapart) is gone, and has been replaced by a global list containing all parts in existance.