Fix the issue with the packet buffer sizes
The problem was caused that async-sending packets after having already managed to send ~some~ data without blocking, and then cancelling that async-send operation resulted in knowledge of the already-sent data to be lost.
As such, Xorg (rightfully so) attempted to re-send data that was already delivered multiple times, with the server receiving that data multiple times as well.