(re-)implement the vmb API for the new mman as mbuilder
The apply function's still kind-of missing, but the complicated part of locking+reflowing mem-parts has already been implemented.
After thinking about it, I've decided to go the same route as the old vm-system did, and provide a builder-system that can be used to implementing exec(). My reasoning for this can be found in "mman-fork.c"