Fix various high-priority bugs with the new mman
It's already able to boot half-way towards mounting the root where it currently fails at an assertion in `mfile_getpart()'. However, following this failure, it _is_ already able to load the builtin debugger and proceed to correctly display tracebacks and all of the usual debug info.

But because it's not 100% working yet (and those 100% may still be a ways off (who knows how many more bugs I'll have to fix...)), the new mman isn't enabled by default, yet.