Fix the bug that caused unwrite not to function
it was actually a problem with `pagedir_syncall()' (which is implected to do a cr3-flip). The problem here was that the __asm__ that did this wasn't __volatile__, and the in/out register operand was left unused, causing gcc to delete the assembly for it. But the most absurd thing about this is the fact that this isn't even a new interface! - pagedir_syncall() is _old_ (but appears to have never actually worked up until now...)